chore: split some e2es
